requirements.txt
# requirements.txt

# File I/O and document processing (from Step 1)
python-docx==1.1.2        # For .docx files
PyPDF2==3.0.1            # For .pdf files
openpyxl==3.1.3          # For .xlsx files
tesseract==0.1.3         # Python wrapper for Tesseract OCR
pytesseract==0.3.13      # Interface to Tesseract for OCR
Pillow==10.4.0           # Image processing for OCR and image extraction
nltk==3.9.1              # For text normalization and sentence tokenization
beautifulsoup4==4.12.3   # For parsing and cleaning HTML content
tqdm==4.66.5             # Progress bars for file processing

# Chunking and tokenization
transformers==4.44.2     # For tokenization (using a pre-trained tokenizer like BERT)
sentence-transformers==3.1.1  # For semantic-aware chunking and future embeddings
regex==2024.9.11         # For advanced regex-based sentence boundary detection
rank_bm25

# Embedding generation
torch==2.4.1             # For GPU acceleration and tensor operations

# FAISS for vector indexing
faiss-cpu==1.9.0         # FAISS for CPU-based vector indexing
joblib==1.4.2            # For parallel processing of domains
pyyaml==6.0.2            # YAML configuration parsing
fire==0.6.0              # For CLI interface
jsonschema==4.23.0       # For configuration validation
concurrent-log-handler==0.9.25  # For thread-safe logging
psutil==6.0.0            # For resource monitoring

#google cloud
google-cloud-storage
dotenv

# API
fastapi
pydantic
uvicorn
python-multipart

# langchian
langchain_community
langchain_huggingface

#MONGODB
pymongo
streamlit